Support Inkscape 1.0.1
This version will run in Inkscape 1.0.1's native version of Python with no changes required.

The bulk of the changes are to be found in the new module fourxidraw_compat.py. New behaviour is triggered via the isPython3() check in that module. See https://wiki.inkscape.org/wiki/index.php/Updating_your_Extension_for_1.0 for the kinds of changes that were necessary.

Main points to note:

1) grbl_serial.py testPort() has code added to cater for serialPort.open() resetting the Arduino. This does happen on my system, but I don't know if it can also happen to systems running older versions of Inkscape. The symptom would be "Error: Unexpected response from GRBL." as a command received an unexpected "Grbl" message.

2) Also in grbl_serial.py, I've had to fix a number of things up to cater for Python3 strings being unicode, and so not directly equivalent to the byte-arrays sent and received by PySerial.

3) all deprecation warnings have been fixed, but Python2 code paths have been kept. Note that I haven't actually verified Python2/Inkscape 0.9.x operation as I don't have this set up.